Add files via upload
This code defines a Company class that stores a collection of employees and provides methods for adding, removing, and printing employees. The Employee class is a base class that defines a print() method which is overridden by the Engineer class to print additional information about their specialization. The code also uses templates to allow the Company class to work with any type of employee.

In the main() function, a Company object is created for Google and two Engineer objects are added to the company. The printEmployees() method is called to print out the employees, and then one employee is removed and the employees are printed again.